adaptive projected guidance

+++ src/diffusers/guiders/classifier_free_zero_star_guidance.py
@@ -97,4 +97,4 @@ def cfg_zero_star_scale(cond: torch.Tensor, uncond: torch.Tensor, eps: float = 1
     squared_norm = torch.sum(uncond**2, dim=1, keepdim=True) + eps
     # st_star = v_cond^T * v_uncond / ||v_uncond||^2
     scale = dot_product / squared_norm
-    return scale.to(cond.dtype)
+    return scale.type_as(cond)
